ah ok so now when I put the barcode in and it doesn;'t search like I asked it for it now wont check alergens without entering ingrediants
How about we have a button for searching for the product to grab the details of that product and then another button to check for alergens?

OR we remove the ingrediant list section and just have a barcode box to search and then we load a seperate page that shows the product with an image and the ingrediants and if it flags for alergens?